+++ plugin/src/main/kotlin/de/infolektuell/gradle/typst/GradleTypstPlugin.kt
@@ -40,6 +40,7 @@ class GradleTypstPlugin : Plugin<Project> {
       project.tasks.withType(TypstCompileTask::class.java).configureEach { task ->
         task.compiler.convention(extension.compiler)
           task.root.convention(project.layout.projectDirectory.asFile.absolutePath)
+          task.useSystemFonts.convention(false)
       }
       extension.sourceSets.all { s ->
           val title = s.name.replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }